Move old data to legacy folder for removal
