+++ src/lib/assets/defaultTestObjects.test.ts
@@ -1,14 +1,18 @@
 import { expect, it } from "@jest/globals";
 import {
-  capabilitiesWithDuplicates,
-  capabilitiesWithFinalize,
-  capabilitiesWithShortKey,
   mockCapabilitiesNew,
-  mockCapabilitiesOld,
   newCapabilityWithDuplicates,
   newCapabilityWithFinalize,
+  newCapabilityWithLongKey,
   newCapabilityWithShortKey,
 } from "./defaultTestObjects";
+import {
+  capabilitiesWithDuplicates,
+  capabilitiesWithFinalize,
+  capabilitiesWithLongKey,
+  capabilitiesWithShortKey,
+  mockCapabilitiesOld,
+} from "./legacy-data/defaultTestObjects";
 
 it("generates the same mockCapability", () => {
   expect(mockCapabilitiesNew).toEqual(mockCapabilitiesOld);
@@ -24,4 +28,8 @@ it("generates the same mockCapabilityWithDuplicates", () => {
 
 it("generates the same mockCapabilityWithShortKey", () => {
   expect(newCapabilityWithShortKey).toEqual(capabilitiesWithShortKey);
+});
+
+it("generates the same mockCapabilityWithLongKey", () => {
+  expect(newCapabilityWithLongKey).toEqual(capabilitiesWithLongKey);
 });
